Django/vjo/requirements.txt
asgiref==3.8.1
basemap==2.0.0
basemap_data==2.0.0
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elementpath==5.0.1
ezgpx==0.2.1
fitparse==1.2.0
folium==0.19.7
fonttools==4.58.4
fpdf2==2.8.3
gmplot==1.4.1
gpxpy==1.6.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pandas==2.3.0
papermap==0.3.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
xmlschema==4.1.0
xyzservices==2025.4.0
